move JndiModel from inline to separate class
It turns out that the approved way of accessing services
in the model is by injecting them in the enclosing
component, then passing them as argument to the model
constructor.

Hmm. Since a large part of service access is in the model,
this limits the advantage of injected components over
using the web application as locator.